# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sparks-Hub Private API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/user/register": {
            "post": {
                "tags": [
                    "User Authentication"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "phone_number",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "text",
                                        "example": "Jane"
                                    },
                                    "last_name": {
                                        "type": "text",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "text",
                                        "example": "example@gmail.com"
                                    },
                                    "phone_number": {
                                        "type": "text",
                                        "example": "+233555555555"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "********"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "example": "**************"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Welcome. You have now successfully signed up for Sparks-Hub"
                                        },
                                        "user": {
                                            "type": "object",
                                            "example": "{id:1,.....}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error during the processing of the request=>Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, registration encountered a problem. please try again later"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "request payload error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/api/v1/user/resend_email_code": {
            "post": {
                "tags": [
                    "User Authentication"
                ],
                "summary": "User email verification code resend",
                "description": "This API is used to a registered user trying to verify their email request for a  new email verification code to be sent to their email",
                "operationId": "UserEmailVerificationCodeResend",
                "parameters": [
                    {
                        "name": "Bearer Token",
                        "in": "header",
                        "description": "Set the user login token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully reset their code",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Your verification code has been sent to your email."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}